src/api-server/requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
pyjwt==2.9.0
bcrypt==4.2.0
python-multipart==0.0.9
requests==2.32.3